Model_Dev_Pipeline/requirements.txt
python-dotenv
youtube_transcript_api
requests
deepeval
huggingface-hub